+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 0.2.1+hotfix.2
+## 0.2.1+3
 
+- Add `auto` attribute whether to play automatically, default is true
+- Optimize the time type
+- Fix repeat callbacks
 - Fix iOS loop playback
 
 ## 0.2.1